Changed player mesh
- Changed player character mesh from the UFO to an animated skeletal
mesh (the default human character in Unreal), for better vision of the
game's looks.
- Needed to change the ColouringBookPawn class to inherit from
ACharacter instead of APawn, so it has further controls on animations
and movement. Renamed the class to ColouringBookCharacter to match this
change.
- Changed the way the inputs are bind for movement, to match the way it
is done in the 3rd person template (the one that uses this skeletal
mesh).
