+++ Source/ColouringBook/ColouringBookCharacter.h
@@ -1,0 +1,57 @@
+// Copyright 1998-2017 Epic Games, Inc. All Rights Reserved.
+#pragma once
+
+#include "GameFramework/Character.h"
+#include "ColouringBookCharacter.generated.h"
+
+UCLASS(Blueprintable)
+class AColouringBookCharacter : public ACharacter
+{
+	GENERATED_BODY()
+
+public:
+	AColouringBookCharacter();
+
+	/** Offset from the ships location to spawn projectiles */
+	UPROPERTY(Category = Gameplay, EditAnywhere, BlueprintReadWrite)
+	FVector GunOffset;
+	
+	/* How fast the weapon will fire */
+	UPROPERTY(Category = Gameplay, EditAnywhere, BlueprintReadWrite)
+	float FireRate;
+
+	/** Sound to play each time we fire */
+	UPROPERTY(Category = Audio, EditAnywhere, BlueprintReadWrite)
+	class USoundBase* FireSound;
+
+	// Begin Actor Interface
+	virtual void Tick(float DeltaSeconds) override;
+	virtual void SetupPlayerInputComponent(class UInputComponent* InputComponent) override;
+	// End Actor Interface
+
+	/* Fire a shot in the specified direction */
+	void FireShot(FVector FireDirection);
+
+	/* Handler for the fire timer expiry */
+	void ShotTimerExpired();
+
+	// Static names for axis bindings
+	static const FName FireForwardBinding;
+	static const FName FireRightBinding;
+
+protected:
+	/** Called for forwards/backward input */
+	void MoveForward(float Value);
+
+	/** Called for side to side input */
+	void MoveRight(float Value);
+
+private:
+
+	/* Flag to control firing  */
+	uint32 bCanFire : 1;
+
+	/** Handle for efficient management of ShotTimerExpired timer */
+	FTimerHandle TimerHandle_ShotTimerExpired;
+};
+
